@@ -1,10 +1,18 @@
 {
   "name": "ts-react-boilerplate",
-  "version": "0.0.1",
+  "version": "0.0.2",
   "description": "A very opinionated (React/TypeScript/Redux/etc) frontend boilerplate",
   "main": "dist/index.js",
   "author": "Lapanti",
   "license": "MIT",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/Lapanti/ts-react-boilerplate.git"
+  },
+  "bugs": {
+    "url": "https://github.com/Lapanti/ts-react-boilerplate/issues"
+  },
+  "homepage": "https://github.com/Lapanti/ts-react-boilerplate#README.md",
   "jest": {
     "snapshotSerializers": [
       "<rootDir>/node_modules/enzyme-to-json/serializer"
@@ -13,7 +21,7 @@
       ".(ts|tsx)": "<rootDir>/node_modules/ts-jest/preprocessor.js"
     },
     "testResultsProcessor": "<rootDir>/node_modules/ts-jest/coverageprocessor.js",
-    "testRegex": "(/__tests__/.*|\\.(test|spec))\\.(ts|tsx)$",
+    "testRegex": "(/__specs__/.*|\\.(spec))\\.(ts|tsx)$",
     "moduleFileExtensions": [
       "ts",
       "tsx",
@@ -34,43 +42,53 @@
     "build:sass": "mkdir -p dist/styles && node-sass src/styles/styles.scss dist/styles/styles.css",
     "build:client": "mkdir -p dist/js && browserify src/index.tsx -p tsify > dist/js/bundle.js",
     "build": "yarn run build:server && yarn run build:sass && yarn run build:client",
-    "start": "NODE_ENV=production node dist/server.js"
+    "start": "NODE_ENV=production node dist/server.js",
+    "docs:clean": "rm -rf _book",
+    "docs:prepare": "gitbook install",
+    "docs:watch": "gitbook serve",
+    "docs:build": "yarn run docs:prepare && gitbook build -g Lapanti/ts-react-boilerplate",
+    "docs:publish": "yarn run docs:clean && yarn run docs:build && cp CNAME _book && cd _book && git init && git commit --allow-empty -m \"Update book\" && git checkout -b gh-pages && touch .nojekyll && git add . && git commit -am \"Update book\" && git push git@github.com:Lapanti/ts-react-boilerplate gh-pages --force"
   },
   "dependencies": {
     "express": "^4.15.2",
-    "react": "^15.4.2",
-    "react-dom": "^15.4.2",
-    "react-redux": "^5.0.3",
+    "react": "^15.5.4",
+    "react-dom": "^15.5.4",
+    "react-redux": "^5.0.4",
     "react-router": "^3.0.2",
     "redux": "^3.6.0",
     "redux-observable": "^0.14.1",
     "rxjs": "^5.2.0"
   },
   "devDependencies": {
-    "@types/es6-shim": "^0.31.32",
     "@types/express": "^4.0.35",
     "@types/jest": "^19.2.2",
     "@types/nock": "^8.2.1",
-    "@types/react": "^15.0.15",
-    "@types/react-dom": "^0.14.23",
-    "@types/react-redux": "^4.4.37",
+    "@types/react": "^15.0.22",
+    "@types/react-dom": "^15.5.0",
+    "@types/react-redux": "^4.4.39",
     "@types/react-router": "^3.0.6",
-    "@types/react-test-renderer": "^15.4.2",
-    "browserify": "^14.1.0",
+    "browserify": "^14.3.0",
     "budo": "^9.4.7",
     "concurrently": "^3.4.0",
     "coveralls": "^2.12.0",
-    "enzyme": "^2.8.0",
-    "enzyme-to-json": "^1.5.0",
+    "enzyme": "^2.8.2",
+    "enzyme-to-json": "^1.5.1",
+    "gitbook-cli": "^2.3.0",
+    "gitbook-plugin-advanced-emoji": "^0.2.2",
+    "gitbook-plugin-anker-enable": "^0.0.4",
+    "gitbook-plugin-edit-link": "^2.0.2",
+    "gitbook-plugin-github": "^3.0.0",
+    "gitbook-plugin-prism": "^2.2.0",
+    "gitbook-plugin-theme-default": "^1.0.6",
     "jest": "^19.0.2",
-    "nock": "^9.0.11",
+    "nock": "^9.0.13",
     "node-sass": "^4.5.2",
-    "react-addons-test-utils": "^15.4.2",
+    "react-test-renderer": "^15.5.4",
     "sass-lint": "^1.10.2",
-    "ts-jest": "^19.0.6",
+    "ts-jest": "^19.0.10",
     "tsify": "^3.0.1",
-    "tslint": "^5.0.0",
-    "tslint-react": "^2.5.0",
+    "tslint": "^5.1.0",
+    "tslint-react": "^2.6.0",
     "typescript": "^2.2.2"
   }
 }
