@@ -1,10 +1,8 @@
 import * as React from 'react';
-import { Dispatch } from 'redux';
-import { Actions } from '../redux/reducer';
 
 export interface IButtonProps {
-    click: () => Dispatch<Actions>;
-    text: string;
+    click(): void;
+    readonly text: string;
 }
 
 const Button: React.StatelessComponent<IButtonProps> = ({ click, text }) => (
