@@ -1,11 +1,9 @@
 import * as React from 'react';
-import { Dispatch } from 'redux';
-import { Actions } from '../redux/reducer';
 import Todo from '../common/Todo';
 
 export interface ITodoComponent {
-    todo: Todo;
-    setDone: (i: number) => Dispatch<Actions>;
+    readonly todo: Todo;
+    setDone(i: number);
 }
 
 const TodoComponent: React.StatelessComponent<ITodoComponent> = ({ todo, setDone }) => (
@@ -14,9 +12,9 @@ const TodoComponent: React.StatelessComponent<ITodoComponent> = ({ todo, setDone
             className="todo__checkbox"
             type="checkbox"
             checked={todo.done}
-            onChange={() => !todo.done && setDone(todo.number)}
+            onChange={() => !todo.done && setDone(todo.id)}
         />
-        <span className="todo__number">{todo.number}:</span>
+        <span className="todo__number">{todo.id}:</span>
         <span className="todo__title">{todo.title}</span>
     </div>
 );
