@@ -1,81 +1,83 @@
 import { ActionsObservable } from 'redux-observable';
-import configureStore from '../../../redux/store';
 import Todo from '../../../common/Todo';
 import IndexReducer, {
     SET_TITLE,
     setTitle,
+    SetTitleAction,
     SAVE_TODO,
     saveTodo,
     SaveTodoAction,
     saveTodoEpic,
     SAVE_TODO_SUCCESS,
     saveTodoSuccess,
+    SaveTodoSuccessAction,
     SET_DONE,
     setDone,
     SetDoneAction,
     setDoneEpic,
     SET_DONE_SUCCESS,
     setDoneSuccess,
+    SetDoneSuccessAction,
     IndexState,
 } from '../IndexReducer';
 
 describe('IndexReducer', () => {
     it('should set the correct title as payload on setTitle', () => {
         const payload = 'THIS_IS_A_TEST_TITLE';
-        const setTitleAction = setTitle(payload);
+        const setTitleAction: SetTitleAction = setTitle(payload);
         expect(setTitleAction).toEqual({ type: SET_TITLE, payload });
-        const newState = IndexReducer(undefined, setTitleAction);
+        const newState: IndexState = IndexReducer(undefined, setTitleAction);
         expect(newState.title).toEqual(payload);
     });
 
     it('should set the correct values on saveTodo', () => {
-        const saveTodoAction = saveTodo();
+        const saveTodoAction: SaveTodoAction = saveTodo();
         expect(saveTodoAction).toEqual({ type: SAVE_TODO });
-        const newState = IndexReducer(undefined, saveTodoAction);
+        const newState: IndexState = IndexReducer(undefined, saveTodoAction);
         expect(newState.loading).toBeTruthy();
     });
 
-    it('should trigger the correct action on saveTodoEpic', async () => {
-        return await saveTodoEpic(ActionsObservable.of(saveTodo()), { getState: () => undefined, dispatch: () => {} })
-            .forEach(actionReceived =>
-                expect(actionReceived)
-                    .toEqual(saveTodoSuccess()));
-    });
+    it('should trigger the correct action on saveTodoEpic', async () => (
+        await saveTodoEpic(ActionsObservable.of(saveTodo()), undefined)
+            .forEach(actionReceived => expect(actionReceived).toEqual({ type: SAVE_TODO_SUCCESS }))
+    ));
 
     it('should set the correct values on saveTodoSuccess', () => {
-        const initialState: IndexState = { title: 'TEST', todos: [], loading: true };
-        const saveTodoSuccessAction = saveTodoSuccess();
+        /* tslint:disable:no-magic-numbers */
+        const testT = new Todo(1, 'Doing', true);
+        const initialState: IndexState = { title: 'TEST', todos: [testT], loading: true };
+        const saveTodoSuccessAction: SaveTodoSuccessAction = saveTodoSuccess();
         expect(saveTodoSuccessAction).toEqual({ type: SAVE_TODO_SUCCESS });
-        const newState = IndexReducer(initialState, saveTodoSuccessAction);
+        const newState: IndexState = IndexReducer(initialState, saveTodoSuccessAction);
         expect(newState.title).toEqual('');
-        expect(newState.todos[0].done).toBeFalsy();
-        expect(newState.todos[0].number).toEqual(1);
-        expect(newState.todos[0].title).toEqual(initialState.title);
+        expect(newState.todos.length).toEqual(2);
+        expect(newState.todos[1].done).toBeFalsy();
+        expect(newState.todos[1].id).toEqual(2);
+        expect(newState.todos[1].title).toEqual(initialState.title);
         expect(newState.loading).toBeFalsy();
+        /* tslint:enable:no-magic-numbers */
     });
 
     it('should set the correct values on setDone', () => {
-        const setDoneAction = setDone(0);
+        const setDoneAction: SetDoneAction = setDone(0);
         expect(setDoneAction).toEqual({ type: SET_DONE, payload: 0 });
-        const newState = IndexReducer(undefined, setDoneAction);
+        const newState: IndexState = IndexReducer(undefined, setDoneAction);
         expect(newState.loading).toBeTruthy();
     });
 
-    it('should trigger the correct action on setDoneEpic', async () => {
-        return await setDoneEpic(ActionsObservable.of(setDone(0)), { getState: () => undefined, dispatch: () => {} })
-            .forEach(actionReceived =>
-                expect(actionReceived)
-                    .toEqual(setDoneSuccess(0)));
-    });
+    it('should trigger the correct action on setDoneEpic', async () => (
+        await setDoneEpic(ActionsObservable.of(setDone(0)), undefined)
+            .forEach(actionReceived => expect(actionReceived).toEqual({ type: SET_DONE_SUCCESS, payload: 0 }))
+    ));
 
     it('should set the correct values on setDoneSuccess', () => {
-        const initialState = { title: '', todos: [new Todo(0, '')], loading: true };
-        const setDoneSuccessAction = setDoneSuccess(0);
+        const initialState: IndexState = { title: '', todos: [new Todo(0, '')], loading: true };
+        const setDoneSuccessAction: SetDoneSuccessAction = setDoneSuccess(0);
         expect(setDoneSuccessAction).toEqual({ type: SET_DONE_SUCCESS, payload: 0 });
-        const newState = IndexReducer(initialState, setDoneSuccessAction);
+        const newState: IndexState = IndexReducer(initialState, setDoneSuccessAction);
         expect(newState.loading).toBeFalsy();
         expect(newState.todos[0].done).toBeTruthy();
-        expect(newState.todos[0].number).toEqual(initialState.todos[0].number);
+        expect(newState.todos[0].id).toEqual(initialState.todos[0].id);
         expect(newState.todos[0].title).toEqual(initialState.todos[0].title);
     });
 });
